--- app/javascript/src/Applications/components/DefaultPlanSelect.scss
@@ -1,5 +1,0 @@
-#default-plan-selector {
-  .pf-c-select__toggle.pf-m-disabled input {
-    background-color: var(--pf-global--disabled-color--300);
-  }
-}
