--- app/javascript/src/Applications/components/DefaultPlanSelectSection.scss
@@ -1,5 +1,0 @@
-@import "~@patternfly/patternfly/components/Card/card.css";
-
-#default-plan-select-card {
-  margin-bottom: var(--pf-c-page__main-section--PaddingBottom);
-}
