+++ app/javascript/src/Applications/components/DefaultPlanSelectWrapper.jsx
@@ -2,11 +2,11 @@
 
 import React from 'react'
 
-import { DefaultPlanSelectSection } from 'Applications'
+import { DefaultPlanSelectCard } from 'Applications'
 import { createReactWrapper } from 'utilities/createReactWrapper'
 
-import type { Props } from 'Applications/components/DefaultPlanSelectSection'
+import type { Props } from 'Applications/components/DefaultPlanSelectCard'
 
-const DefaultPlanSelectWrapper = (props: Props, containerId: string) => createReactWrapper(<DefaultPlanSelectSection {...props} />, containerId)
+const DefaultPlanSelectWrapper = (props: Props, containerId: string) => createReactWrapper(<DefaultPlanSelectCard {...props} />, containerId)
 
 export { DefaultPlanSelectWrapper }
