+++ app/javascript/src/Applications/index.js
@@ -1,5 +1,5 @@
 // @flow
 
 export * from 'Applications/components/DefaultPlanSelect'
-export * from 'Applications/components/DefaultPlanSelectSection'
+export * from 'Applications/components/DefaultPlanSelectCard'
 export * from 'Applications/components/DefaultPlanSelectWrapper'
